Grpc gateway merge

+++ test/globalSetup.ts
@@ -4,11 +4,15 @@ import { exec, sleep } from "./utils";
 require("ts-node").register({ transpileOnly: true });
 
 module.exports = async () => {
+  if (process.env.SKIP_LOCALSECRET === "true") {
+    return;
+  }
+
   // init localsecret
   console.log("\nSetting up LocalSecret...");
   await exec("docker rm -f localsecret || true");
   const { /* stdout, */ stderr } = await exec(
-    "docker run -it -d -p 9091:9091 --name localsecret ghcr.io/scrtlabs/localsecret:v1.5.0-beta.11",
+    "docker run -it -d -p 1317:1316 --name localsecret ghcr.io/scrtlabs/localsecret:v1.6.0-alpha.4",
   );
 
   // console.log("stdout (testnet container id?):", stdout);
@@ -35,8 +39,8 @@ module.exports = async () => {
 
 async function waitForBlocks() {
   while (true) {
-    const secretjs = await SecretNetworkClient.create({
-      grpcWebUrl: "http://localhost:9091",
+    const secretjs = new SecretNetworkClient({
+      url: "http://localhost:1317",
       chainId: "secretdev-1",
     });
 
@@ -47,7 +51,7 @@ async function waitForBlocks() {
         break;
       }
     } catch (e) {
-      // console.error(e);
+      // console.eerror(e);
     }
     await sleep(250);
   }
